WIP: SASS-5681: Build failing due to coverage. Reorganising tests
