+++ test/controllers/OtherIncomeControllerSpec.scala
@@ -1,19 +1,16 @@
 package controllers
 
 import base.SpecBase
-import forms.OtherIncomeFormProvider
 import models.{NormalMode, UserAnswers}
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.OtherIncomePage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import repositories.SessionRepository
-import views.html.OtherIncomeView
 
 import scala.concurrent.Future
